libcollectdclient: Separate module for lcc_network_parse().
